Split the LSP features into separate modules
