+++ src/adapters/jupyterlab/components/free_tooltip.ts
